restructure exceptions module
