--- darwin/future/exceptions/client.py
@@ -1,9 +1,0 @@
-from darwin.future.exceptions.base import DarwinException
-
-
-class NotFound(DarwinException):
-    pass
-
-
-class Unauthorized(DarwinException):
-    pass
