--- darwin/future/exceptions/files.py
@@ -1,5 +1,0 @@
-from darwin.future.exceptions.base import DarwinException
-
-
-class UnrecognizableFileEncoding(DarwinException):
-    pass
