--- darwin/future/exceptions/meta.py
@@ -1,13 +1,0 @@
-from darwin.future.exceptions.base import DarwinException
-
-
-class MetaException(DarwinException):
-    pass
-
-
-class MissingSlug(MetaException):
-    pass
-
-
-class MissingDataset(MetaException):
-    pass
