--- darwin/future/exceptions/query.py
@@ -1,17 +1,0 @@
-from darwin.future.exceptions.base import DarwinException
-
-
-class ResultsNotFound(DarwinException):
-    pass
-
-
-class MoreThanOneResultFound(DarwinException):
-    pass
-
-
-class InvalidQueryModifier(DarwinException):
-    pass
-
-
-class InvalidQueryFilter(DarwinException):
-    pass
